\name{onlinePCA-package}
\alias{onlinePCA-package}
\alias{onlinePCA}
\docType{package}
\title{
\packageTitle{onlinePCA}
}
\description{
\packageDescription{onlinePCA}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{onlinePCA}
\packageIndices{onlinePCA}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{onlinePCA}

Maintainer: \packageMaintainer{onlinePCA}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
