\name{onlinePCA-package}
\alias{onlinePCA-package}
\alias{onlinePCA}
\docType{package}
\title{
Online Principal Component Analysis
}
\description{
This package implements two types of online PCA algorithms, recursive and stochastic gradient, as well as convenience functions for updating the sample mean and sample covariance.}

\details{
\tabular{ll}{
Package: \tab onlinePCA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-05-07\cr
License: \tab GPL-3\cr
}
%~~ An overview of how to use the package, including the most important functions ~~
}
\author{
David Degras <ddegrasv@depaul.edu>
}
\references{
Gu, M. and Eisenstat, S. C. (1994). A stable and efficient algorithm for the rank-one modification of the symmetric eigenproblem. \emph{SIAM Journal of Matrix Analysis and Applications.}\cr
Hegde et al. (2006) Perturbation-Based Eigenvector Updates for On-Line Principal Components Analysis and Canonical Correlation Analysis. \emph{Journal of VLSI Signal Processing}. \cr
Li, W., Yue, H. H., Valles-Cervantes, S. and Qin, S. J. (2000). Recursive PCA for adaptive process monitoring. \emph{Journal of Process Control}.\cr
Oja (1992). Principal components, Minor components, and linear neural networks. \emph{Neural Networks.} 
}
% \keyword{ package }

